requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.4
django-ckeditor==5.4.0
django-cors-headers==2.2.0
django-filter==1.1.0
django-js-asset==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
djoser==1.1.5
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.1.0
Pillow==5.0.0
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2017.3
requests==2.19.0
requests-oauthlib==1.0.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
